created ImagEditor component
